// apps/crypto-mobile/src/interfaces/TickerResponse.ts
export interface TicketResponse {
  best_ask: string;
  best_ask_size: string;
  best_bid: string;
  best_bid_size: string;
  high_24h: string;
  last_size: string;
  low_24h: string;
  open_24h: string;
  price: string;
  product_id: string;
  sequence: number;
  side: string;
  time: string;
  trade_id: number;
  type: string;
  volume_24h: string;
  volume_30d: string;
}